See changelog 2024-07-12

+++ app/internal/calculations/algo/section/pub/rect/cutting/entity/cutting.go
@@ -24,6 +24,6 @@ type (
 
 	// QuantityResult - результат работы алгоритма AlgoQuantity.
 	QuantityResult struct {
-		Quantity int32 `json:"quantity"`
+		Quantity uint64 `json:"quantity"`
 	}
 )
